+++ InheritanceProject1/InheritanceProject1/Models/Laptop.cs
@@ -8,7 +8,7 @@ namespace InheritanceProject1.Models
 {
     public class Laptop : Computer
     {
-        public Laptop(string manufacturer, int cpuMhz, int cpuCores) : base(manufacturer, cpuMhz, cpuCores)
+        public Laptop(string manufacturer, float cpuMhz, int cpuCores) : base(manufacturer, cpuMhz, cpuCores)
         {
 
         }
